% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{bdc_parser}
\alias{bdc_parser}
\title{Brazil Data Cube File Parser Metadata}
\usage{
bdc_parser()
}
\description{
Example parser for extracting information from files used for
generating the \code{odc_cube}. This is the default parser used by
\code{odc-sits} when the data is loaded.
}
