% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble.R
\name{.build_odcsits_tb}
\alias{.build_odcsits_tb}
\title{Build odc-sits tibble}
\usage{
.build_odcsits_tb(datasets, datasets_files)
}
\arguments{
\item{datasets}{a \code{data.frame} from Open Data Cube Database}

\item{datasets_files}{a \code{data.frame} from Open Data Cube Database}
}
\value{
a \code{tibble} with database data.
}
\description{
Creates a list of the odcsits class. The "dataset" and
"dataset_files" information is transformed into \code{tibble::tibble}.
}
