% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odc_index.R
\name{metadata query view}
\alias{metadata query view}
\alias{.odc_query_view}
\title{Metadata View}
\usage{
.odc_query_view()
}
\description{
This function creates a temporary database view in the
Open Data Cube that makes it easy to search for other operations.
In addition, the created view provides spatio-temporal, bands, and CRS
information of the indexed datasets.
}
