% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odc_cube.R
\name{odc_brick}
\alias{odc_brick}
\title{ODC Brick Cube}
\usage{
odc_brick(odc_index, satellite, sensor, datasets, ...)
}
\arguments{
\item{odc_index}{...}

\item{satellite}{...}

\item{datasets}{...}

\item{...}{...}
}
\description{
...
}
