% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odc_cube.R
\name{odc_cube}
\alias{odc_cube}
\title{Open Data Cube data}
\usage{
odc_cube(satellite, sensor, datasets, .parser = bdc_parser())
}
\arguments{
\item{satellite}{Satellite that produced the images. This name must match
the satellites that are declared in the \code{sits} package
configuration file. (see details below)}

\item{sensor}{Sensor that produced the images. (see details below)}

\item{datasets}{a \code{tibble} object from class \code{odcsits} with the
datasets that will be loaded from ODC instance to build
the data cube.}

\item{.parser}{a \code{list} with parsing information for files.}
}
\description{
Data cube created with data indexed in an Open Data Cube
instance. The construction of the data cube is based on the data structures
in the \code{sits} package.
}
\details{
The "satellite" and "sensor" parameters must match the declarations
made in the \code{sits} package configuration file.

The ".parser" it is an optional parameter that can be used to declare
the naming pattern of the files indexed by the Open Data Cube. This pattern
will be used to define the characteristics of the data cube that will
be generated. For this parameter a named \code{vector} ("delim", "parse_info")
is expected. The value of "delim" is used to determine the separator in the
name of the files loaded from the ODC. On the other hand, the "parse_info"
parameter determines what information is represented by each position in the
configuration file. For example, for the file name:

   CB4_64_16D_STK_v001_022024_2019-10-16_2019-10-31_BAND15

The ".parser" parameter will be:

.parser = list(
 delim = "_",
 parse_info = c("sat", "res", "X1", "X2", "X3", "tile", "date", "X4", "band")
)

This will split the file name using "_" and the positions have the meaning
 represented in "parse_info". In this case:

  sat = CB4
  res = 64
  tile = 022024
  date = 2019-10-16_2019-10-31
  band = BAND15

The "Xx" values represent values that should not be used.
}
\note{
It is assumed that the data loaded by the \code{odc_cube} function is
aligned in time and space. If the data does not have these properties,
the function may have problems.
}
