% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odc_data.R
\name{odc_products}
\alias{odc_products}
\title{Open Data Cube Products}
\usage{
odc_products(index, ...)
}
\arguments{
\item{index}{a \code{tibble} object from class \code{odcsits-index} with
Open Data Cube Database valid connection.}

\item{...}{parameters passed to \code{DBI::dbGetQuery} function to
control the connection made with Open Data Cube Database}
}
\description{
Lists the products that are available for use in the queried
Open Data Cube instance.
}
