% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odc_data.R
\name{odc_search}
\alias{odc_search}
\title{Open Data Cube Datasets Search}
\usage{
odc_search(index, product, start_date, end_date, bbox = NULL, ...)
}
\arguments{
\item{index}{a \code{tibble} object from class \code{odcsits-index} with
Open Data Cube Database valid connection.}

\item{product}{Name of the product that will be searched in the Open Data Cube.}

\item{start_date}{Initial date for the search data}

\item{end_date}{Final date for the search data}

\item{bbox}{Area of interest (see details below)}

\item{...}{parameters passed to \code{DBI::dbGetQuery} function to
control the connection made with Open Data Cube Database}
}
\description{
Searches datasets of a Product in an Open Data Cube instance.
The search can consider both spatial and temporal information.
}
\details{
The "bbox" parameters allows a selection of an area of interest.
Either using a named \code{vector} ("xmin", "ymin", "xmax", "ymax") with
values in WGS 84.
}
