% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odc_cube.R, R/odc_data.R
\name{odc_cube}
\alias{odc_cube}
\alias{odc_products}
\alias{odc_search}
\title{ODC Complete Cube}
\usage{
odc_cube(odc_index, satellite, sensor, datasets)

odc_products(odc_index)

odc_search(odc_index, collection, start_date, end_date, bbox = NULL)
}
\arguments{
\item{odc_index}{...}

\item{satellite}{...}

\item{sensor}{...}

\item{datasets}{...}

\item{collection}{...}

\item{start_date}{...}

\item{end_date}{...}

\item{bbox}{...}
}
\description{
...

...

...
}
