% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odc_index.R
\name{odc_index}
\alias{odc_index}
\title{ODC Products Search}
\usage{
odc_index(dbname, host, port, user, password)
}
\arguments{
\item{dbname}{...}

\item{host}{...}

\item{port}{...}

\item{user}{...}

\item{password}{...}
}
\description{
...
}
